+++ src/TestDiv.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const TestDiv = () => {
+	return <div className='tw-m-10 tw-bg-amber-600 tw-p-5'>HI there</div>
+}
+export default TestDiv