+++ src/index.js
@@ -6,9 +6,21 @@ import SingleDatePicker from './SingleDatePicker';
 import TimePicker from './TimePicker';
 import ScrollTimePicker from './ScrollTimePicker';
 
+import DateRangePickerTW from './Tailwind/DateRangePickerTW';
+import SingleDatePickerTW from './Tailwind/SingleDatePickerTW';
+import TimePickerTW from './Tailwind/TimePickerTW';
+
+import TestDiv from './TestDiv';
+
 export {
   DateRangePicker,
   SingleDatePicker,
 	TimePicker,
 	ScrollTimePicker,
+
+	DateRangePickerTW,
+	SingleDatePickerTW,
+	TimePickerTW,
+
+	TestDiv
 }